pyproject.toml
[project]
name = "n8n-cli"
version = "0.1.0"
description = "CLI and Python client for the n8n Public API"
requires-python = ">=3.13"
dependencies = [
    "httpx>=0.27.0",
    "attrs>=24.0.0",
    "python-dateutil>=2.9.0",
    "typer>=0.15.0",
    "rich>=13.0.0",
]

[project.scripts]
n8n = "n8n_cli.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = ["F", "I", "UP"]
